requirements.txt
pyautogui
pycaw
psutil
playsound
Pillow
comtypes
pywin32

